services/graphics_service/requirements.txt
torch
diffusers
transformers
accelerate
open-clip-torch
psutil
Pillow

